Minor refactor / move components views to subfolder Rename uiIDs to app under const folder Move 'back to top' to global app.js Add dynamic app titles using DocumentTitle
