+++ src/scripts/app.js
@@ -1,14 +1,15 @@
 /** @jsx React.DOM */
 var React = require('react/addons');
 var Router = require('react-router');
+var jQueryScrollTo = require('jquery.scrollto');
+
+var AppConsts = require('./consts/app');
 
 /* Load views */
-//var Navigation = require('./views/navigation');
-//var PortfolioList = require('./views/portfolioList');
 var Home = require("./views/home");
-var PortfolioDetail = require('./views/portfolioDetail');
+var PortfolioDetail = require('./views/detail');
 var NotFound = require("./views/404");
-var LoadingStatus = require('./views/loadingStatus');
+var LoadingStatus = require('./views/components/loadingStatus');
 var ReactCSSTransitionGroup = require('react/lib/ReactCSSTransitionGroup');
 
 /* Load Router */
@@ -19,6 +20,16 @@ var RouteHandler = Router.RouteHandler;
 
 var App = React.createClass({
 
+  backToTop: function(e){
+
+    e.preventDefault();
+
+    window.setTimeout(function(){
+
+      jQueryScrollTo('#' + AppConsts.UIID.masthead, 500)
+    },0);
+  },
+
   render: function(){
 
     var route = window.location.hash;
@@ -27,6 +38,9 @@ var App = React.createClass({
       <LoadingStatus />
       <ReactCSSTransitionGroup transitionName="example">
         <RouteHandler key={route} />
+        <div className="text-center mtl">
+          <a href="#" onClick={this.backToTop}>Back to top</a>
+        </div>
       </ReactCSSTransitionGroup>
     </div>);
   }
