+++ src/scripts/consts/app.js
@@ -1,0 +1,8 @@
+module.exports = {
+  TITLE: 'Syn By Design - Eric Masiello\'s portfolio',
+  UIID: {
+    masthead: 'masthead',
+    about: 'about',
+    portfolioList: 'portfolioList'
+  }
+};
