--- src/scripts/props/uiIDs.js
@@ -1,6 +1,0 @@
-module.exports = {
-
-  masthead: 'masthead',
-  about: 'about',
-  portfolioList: 'portfolioList'
-};
