+++ src/scripts/views/404.js
@@ -1,8 +1,14 @@
 var React = require('react/addons');
+var DocumentTitle = require('react-document-title');
+var AppConsts = require('../consts/app');
 
 var NotFound = React.createClass({
     render: function () {
-        return <h2>Not found</h2>;
+        return (
+          <DocumentTitle title={'404 Page Not Found - ' + AppConsts.TITLE}>
+            <h2>Not found</h2>
+          </DocumentTitle>
+          );
     }
 });
