+++ src/scripts/views/components/navigation.js
@@ -1,7 +1,7 @@
 var React = require('react/addons');
 var Link = require('react-router').Link;
 var jQueryScrollTo = require('jquery.scrollto');
-var UIIDs = require('../props/uiIDs');
+var AppConsts = require('../../consts/app');
 
 var Navigation = React.createClass({
 
@@ -15,17 +15,17 @@ var Navigation = React.createClass({
 
   render: function(){
 
-    var logo = '<use xlink:href="#syn-logo" />';
-    var title = '<use xlink:href="#syn-title" />';
-    var slogan = '<use xlink:href="#syn-slogan" />';
+    var svgLogo = '<use xlink:href="#syn-logo" />';
+    var svgTitle = '<use xlink:href="#syn-title" />';
+    var svgSlogan = '<use xlink:href="#syn-slogan" />';
 
     var logo = (
       <li role="presentation" className="col-xs-12  col-sm  col-md-5  first-xs  default-sm  masthead__logo">
         <h1 className="masthead__logo__title">
           <Link to="home">
-            <svg className="logo  logo__head" dangerouslySetInnerHTML={{__html:logo}} />
-            <svg className="logo  logo__title" dangerouslySetInnerHTML={{__html:title}} />
-            <svg className="logo  logo__slogan" dangerouslySetInnerHTML={{__html:slogan}} />
+            <svg className="logo  logo__head" dangerouslySetInnerHTML={{__html:svgLogo}} />
+            <svg className="logo  logo__title" dangerouslySetInnerHTML={{__html:svgTitle}} />
+            <svg className="logo  logo__slogan" dangerouslySetInnerHTML={{__html:svgSlogan}} />
           </Link>
         </h1>
       </li>
@@ -34,10 +34,10 @@ var Navigation = React.createClass({
     var nav = (
       <ul className="row  flex-center  middle-xs  masthead__nav">
         <li role="presentation" className="col-xs-12  col-sm">
-          <Link to="home" className="pill" onClick={this.scrollTo.bind(this, UIIDs['about'])}>About</Link>
+          <Link to="home" className="pill" onClick={this.scrollTo.bind(this, AppConsts.UIID.about)}>About</Link>
         </li>
         <li role="presentation" className="col-xs-12  col-sm">
-          <Link to="home" className="pill" onClick={this.scrollTo.bind(this, UIIDs['portfolioList'])}>Portfolio</Link>
+          <Link to="home" className="pill" onClick={this.scrollTo.bind(this, AppConsts.UIID.portfolioList)}>Portfolio</Link>
         </li>
         {logo}
         <li role="presentation" className="col-xs-12  col-sm"><a href="https://www.linkedin.com/in/ericmasiello" className="pill">LinkedIn</a></li>
@@ -67,7 +67,7 @@ var Navigation = React.createClass({
     }
 
     return (
-      <header className="header  masthead" id={UIIDs.masthead}>
+      <header className="header  masthead" id={AppConsts.UIID.masthead}>
         <div className="container-fluid">
           <div className="row  masthead__decoration-container  middle-xs">
             <div className="col-xs  masthead__decoration-line"></div>
