+++ src/scripts/views/components/portfolioItem.js
